+++ qiskit_runtime/__init__.py
@@ -11,5 +11,3 @@
 # that they have been altered from the originals.
 
 """Main entry point for the qiskit_runtime"""
-
-from .result import RuntimeResult
